requirements.txt
# --- MCP Framework ---
mcp[cli]>=1.9.0

# --- Configuration ---
pydantic-settings>=2.0

# --- S3 (Token Store + Vault Storage Sync) ---
boto3>=1.35.0

# --- OpenBao client (compatible Vault API) ---
hvac>=2.3.0

# --- HTTP client (health checks, CLI) ---
httpx>=0.27.0

# --- ASGI ---
uvicorn[standard]>=0.32.0

# --- CLI (scripts/mcp_cli.py) ---
click>=8.1
rich>=13.0
prompt-toolkit>=3.0
python-dotenv>=1.0

# --- Tests (intégration réelle, pas de mocking) ---
pytest>=8.0
pytest-asyncio>=0.24.0
